pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "conviction-gradient-framework"
version = "1.0.0"
description = "Gradient-descent conviction engine for L/S portfolio sizing, IC governance, and AI evaluation"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "bds-data-providers @ git+https://github.com/bdschi1/financial-data-providers.git",
    "polars>=1.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "yfinance>=0.2.30",
]

[project.optional-dependencies]
api = ["fastapi>=0.109", "uvicorn>=0.25"]
bloomberg = ["blpapi>=3.19"]
ibkr = ["ib_insync>=0.9.86"]
dev = ["pytest>=8.0", "pytest-cov>=6.0", "ruff>=0.5", "httpx>=0.27"]
all = [
    "fastapi>=0.109",
    "uvicorn>=0.25",
]

[project.scripts]
cgf = "cli:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [
    "engine",
    "components",
    "sizing",
    "governance",
    "evaluation",
    "bridges",
    "storage",
    "config",
    "api",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = [
    "engine",
    "components",
    "sizing",
    "governance",
    "evaluation",
    "bridges",
    "storage",
    "config",
    "api",
]
